% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_armadillo.R
\name{du.armadillo.create.project}
\alias{du.armadillo.create.project}
\title{Create projects in Armadillo}
\usage{
du.armadillo.create.project(project)
}
\arguments{
\item{project}{the target project where the data should be put into}
}
\description{
Create projects in Armadillo
}
\keyword{internal}
