% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_armadillo.R
\name{du.armadillo.import}
\alias{du.armadillo.import}
\title{Import data into Armadillo}
\usage{
du.armadillo.import(
  project,
  data,
  dict_version,
  dict_kind,
  data_version,
  table_type
)
}
\arguments{
\item{data}{data to upload}

\item{dict_version}{data model version}

\item{dict_kind}{data model version}

\item{data_version}{data model version}

\item{table_type}{data model version}
}
\description{
Import data into Armadillo
}
\keyword{internal}
