% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_armadillo.R
\name{du.armadillo.login}
\alias{du.armadillo.login}
\title{Login to the Armadillo backend}
\usage{
du.armadillo.login(login_data)
}
\arguments{
\item{login_data}{contains the server credentials to authenticate in Armadillo}
}
\description{
Login to the Armadillo backend
}
\keyword{internal}
