% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{du.check.action}
\alias{du.check.action}
\title{Check is action is the correct value}
\usage{
du.check.action(action = "all")
}
\arguments{
\item{action}{action to perform}
}
\description{
Check is action is the correct value
}
\keyword{internal}
