% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{du.check.session}
\alias{du.check.session}
\title{Check if there is an active session with a DataSHIELD backend}
\usage{
du.check.session(upload = FALSE)
}
\arguments{
\item{upload}{is a session needed or not}
}
\description{
Check if there is an active session with a DataSHIELD backend
}
\keyword{internal}
