% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_helpers.R
\name{du.check.variables}
\alias{du.check.variables}
\title{Check if there are columns not matching the dictionary.}
\usage{
du.check.variables(dict_kind, data_columns, non_interactive)
}
\arguments{
\item{dict_kind}{specify which dictionary you want to check}

\item{data_columns}{the coiumns within the data}

\item{non_interactive}{if set to TRUE you will get no questions}
}
\value{
stops the program if someone terminates
}
\description{
Check if there are columns not matching the dictionary.
}
\keyword{internal}
