% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{du.check.version}
\alias{du.check.version}
\title{Check if the given version matches the syntax number . underscore . number}
\usage{
du.check.version(version)
}
\arguments{
\item{version}{the version input of the user}
}
\description{
Check if the given version matches the syntax number . underscore . number
}
\keyword{internal}
