% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{du.clean.temp.workdir}
\alias{du.clean.temp.workdir}
\title{Cleanup all the temporary files from the upload}
\usage{
du.clean.temp.workdir(upload, workdirs)
}
\arguments{
\item{upload}{should we upload the contents to the backend}

\item{workdirs}{a list containing the original workdir and the created workdir}
}
\description{
Cleanup all the temporary files from the upload
}
\keyword{internal}
