% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{du.create.temp.workdir}
\alias{du.create.temp.workdir}
\title{Create a temporary directory in the current working directory to store all temporary files}
\usage{
du.create.temp.workdir()
}
\description{
Create a temporary directory in the current working directory to store all temporary files
}
\keyword{internal}
