% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_helpers.R
\name{du.data.frame.remove.all.na.rows}
\alias{du.data.frame.remove.all.na.rows}
\title{Get the table without rows containing only NA's.}
\usage{
du.data.frame.remove.all.na.rows(dataframe)
}
\arguments{
\item{dataframe}{dataframe to check}
}
\value{
dataframe without the na values
}
\description{
We have to remove the first column (child_id), that is generated always.
}
\keyword{internal}
