% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{du.dict.import}
\alias{du.dict.import}
\title{Import the data dictionaries into Opal}
\usage{
du.dict.import(project, dictionaries, data_version)
}
\arguments{
\item{project}{project in which the data is imported}

\item{dictionaries}{all the dictionaries pulled from the repository}

\item{data_version}{data version to put into the table}
}
\description{
Import the data dictionaries into Opal
}
\keyword{internal}
