% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{du.dict.retrieve.tables}
\alias{du.dict.retrieve.tables}
\title{Retrieve the released dictionaries from 'ds-dictionaries' to match against}
\usage{
du.dict.retrieve.tables(api_url, dict_name, dict_version, data_version)
}
\arguments{
\item{api_url}{url to retrieve the tables from}

\item{dict_version}{model version of the data}

\item{data_version}{data version if used to create the tables}

\item{dict_kind}{dictionary kind can be outcome, core, exposure}
}
\value{
dictionaries list of tables
}
\description{
Retrieve the released dictionaries from 'ds-dictionaries' to match against
}
\keyword{internal}
