% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{du.get.response.as.dataframe}
\alias{du.get.response.as.dataframe}
\title{Parse response to dataframe}
\usage{
du.get.response.as.dataframe(url)
}
\arguments{
\item{url}{get this location}
}
\value{
response as dataframe
}
\description{
Parse response to dataframe
}
\keyword{internal}
