% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{du.login}
\alias{du.login}
\title{Login into the DataSHIELD backend}
\usage{
du.login(login_data)
}
\arguments{
\item{login_data}{login data frame containing the server url}
}
\description{
Login into the DataSHIELD backend
}
\examples{
\dontrun{

login_data <- data.frame(
  server = "https://armadillo.dev.molgenis.org",
  storage = "https://armadillo-minio.dev.molgenis.org",
  username = "admin",
  password = "admin",
  insecure = FALSE,
  options = NULL
)

du.login(login_data)
}
}
