% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_helpers.R
\name{du.match.columns}
\alias{du.match.columns}
\title{Matched the columns in the source data.
You can then match the found column against the dictionary.}
\usage{
du.match.columns(data_columns, dict_columns)
}
\arguments{
\item{data_columns}{columns obtained from raw data}

\item{dict_columns}{columns matched in the dictionary}
}
\value{
matched_columns in source data
}
\description{
Matched the columns in the source data.
You can then match the found column against the dictionary.
}
\keyword{internal}
