% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{du.match.dict.categories}
\alias{du.match.dict.categories}
\title{Match the categories with the variables to be import them
Import the variables}
\usage{
du.match.dict.categories(project, dict_kind, table)
}
\arguments{
\item{project}{project resource in Opal}

\item{table}{dictionary to upload to Opal}

\item{source_file}{source file for the dictionaries}
}
\description{
Match the categories with the variables to be import them
Import the variables
}
\keyword{internal}
