% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_clocks.R
\name{du.methyl.clocks}
\alias{du.methyl.clocks}
\title{Generate clocks for uploading to Opal}
\usage{
du.methyl.clocks(methylation_data_input_path, header_data_input_path)
}
\arguments{
\item{methylation_data_input_path}{path to the methylation data}

\item{header_data_input_path}{path to the data that binds it to child_id's}
}
\description{
Generate clocks for uploading to Opal
}
\examples{
 \dontrun{
du.clocks(
  methylation_data_input_path = ~/path-to-file/methylation_data_for_all_children.csv,
  header_data_input_path = "~/path-to-file/children_header_data"
)
}

}
