% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{du.num.extract}
\alias{du.num.extract}
\title{Numerical extraction function
Number at the end of the string: Indicates year. We need to extract this to create the age_years variable.
This is the function to do so.}
\usage{
du.num.extract(input_string)
}
\arguments{
\item{input_string}{convert this string into an integer value}
}
\description{
Numerical extraction function
Number at the end of the string: Indicates year. We need to extract this to create the age_years variable.
This is the function to do so.
}
\keyword{internal}
