% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_opal.R
\name{du.opal.data.import}
\alias{du.opal.data.import}
\title{Importing generated data files}
\usage{
du.opal.data.import(project, dict_kind, file_name)
}
\arguments{
\item{project}{project to import the data into}

\item{dict_kind}{can be 'core' or 'outcome'}

\item{file_name}{name of the data file}
}
\description{
Importing generated data files
}
\keyword{internal}
