% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_opal.R
\name{du.opal.login}
\alias{du.opal.login}
\title{Opal login}
\usage{
du.opal.login(login_data)
}
\arguments{
\item{login_data}{data frame containing login credentials}
}
\description{
Opal login
}
\keyword{internal}
