% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_opal.R
\name{du.opal.upload}
\alias{du.opal.upload}
\title{Uploading the generated data files to Opal backend}
\usage{
du.opal.upload(dict_kind, file_name)
}
\arguments{
\item{dict_kind}{can be 'core' or 'outcome'}

\item{file_name}{name of the data file}
}
\description{
Uploading the generated data files to Opal backend
}
\keyword{internal}
