% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{du.populate.beta}
\alias{du.populate.beta}
\title{Create tables in Opal to import the data for the beta dictionaries}
\usage{
du.populate.beta(dict_name, database_name)
}
\arguments{
\item{dict_name}{dictionary path to search on}

\item{database_name}{name of the database in Opal}
}
\description{
Create tables in Opal to import the data for the beta dictionaries
}
\keyword{internal}
