% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{du.populate.dict.versions}
\alias{du.populate.dict.versions}
\title{Get the possible dictionary versions from Github}
\usage{
du.populate.dict.versions(dict_kind, dict_version)
}
\arguments{
\item{dict_kind}{dictionary kind (can be 'core' or 'outcome')}

\item{dict_version}{dictionary version (can be 'x_x')}
}
\description{
Get the possible dictionary versions from Github
}
\keyword{internal}
