% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_helpers.R
\name{du.populate.match.categories}
\alias{du.populate.match.categories}
\title{Match the categories with the variables to be import them
Import the variables}
\usage{
du.populate.match.categories(
  project,
  table,
  variables,
  categories,
  source_file
)
}
\arguments{
\item{project}{project resource in Opal}

\item{table}{dictionary to upload to Opal}

\item{variables}{dictionary variables to upload}

\item{categories}{dictionary categories to upload}

\item{source_file}{source file for the dictionaries}
}
\description{
Match the categories with the variables to be import them
Import the variables
}
