% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{du.populate}
\alias{du.populate}
\title{Populate your Opal instance with the new version of the data dictionary
Involves only the core variables}
\usage{
du.populate(dict_version, cohort_id, data_version, database_name, dict_kind)
}
\arguments{
\item{dict_version}{dictionary version}

\item{cohort_id}{cohort identifier (possible values are: 'dnbc', 'gecko', 'alspac', 'genr', 'moba', 'sws', 'bib', 'chop', 'elfe', 'eden', 'ninfea', 'hbcs', 'inma', 'isglobal', 'nfbc66', 'nfbc86', 'raine', 'rhea')}

\item{data_version}{version of the data (specific to the cohort)}

\item{database_name}{the database name specified in your Opal instance (defaults to 'opal_data')}

\item{dict_kind}{dictionnary kind, can be 'core' or 'outcome'}
}
\description{
Populate your Opal instance with the new version of the data dictionary
Involves only the core variables
}
\keyword{internal}
