% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{du.project.create}
\alias{du.project.create}
\title{Create the project with data dictionary version in between}
\usage{
du.project.create(project, database_name)
}
\arguments{
\item{project}{prpject resource in Opal}

\item{database_name}{the database name of the Opal instance (default = opal_data)}
}
\description{
Create the project with data dictionary version in between
}
\keyword{internal}
