% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.child.birth.out}
\alias{du.quality.local.core.child.birth.out}
\title{Part of child characteristics
Child birth outcomes
Involves variables 79-97}
\usage{
du.quality.local.core.child.birth.out(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of child characteristics
Child birth outcomes
Involves variables 79-97
}
\keyword{internal}
