% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.child.exposure.lifestyle.environ.char}
\alias{du.quality.local.core.child.exposure.lifestyle.environ.char}
\title{Part of child characteristics
Child exposures/lifestyle/environment
Involves variables 104-115}
\usage{
du.quality.local.core.child.exposure.lifestyle.environ.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of child characteristics
Child exposures/lifestyle/environment
Involves variables 104-115
}
\keyword{internal}
