% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.child.health.related.char}
\alias{du.quality.local.core.child.health.related.char}
\title{Part of child characteristics
Child health-related characteristics
Involves variables 98-103}
\usage{
du.quality.local.core.child.health.related.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of child characteristics
Child health-related characteristics
Involves variables 98-103
}
\keyword{internal}
