% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.household.char}
\alias{du.quality.local.core.household.char}
\title{Household characteristics
Involves variables 116-115}
\usage{
du.quality.local.core.household.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Household characteristics
Involves variables 116-115
}
\keyword{internal}
