% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.maternal.health.char}
\alias{du.quality.local.core.maternal.health.char}
\title{Part of maternal characteristics
Health related characteristics
This involes variables 18-31}
\usage{
du.quality.local.core.maternal.health.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of maternal characteristics
Health related characteristics
This involes variables 18-31
}
\keyword{internal}
