% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.maternal.lifestyle.char}
\alias{du.quality.local.core.maternal.lifestyle.char}
\title{Part of maternal characteristics
Lifestyle charasteristics
This involves variables 32-38}
\usage{
du.quality.local.core.maternal.lifestyle.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of maternal characteristics
Lifestyle charasteristics
This involves variables 32-38
}
\keyword{internal}
