% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.maternal.obstetric.char}
\alias{du.quality.local.core.maternal.obstetric.char}
\title{Part of maternal characteristics
Obstetric charasteristics
This involves variables 39-45}
\usage{
du.quality.local.core.maternal.obstetric.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of maternal characteristics
Obstetric charasteristics
This involves variables 39-45
}
\keyword{internal}
