% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.maternal.social.char}
\alias{du.quality.local.core.maternal.social.char}
\title{Part of maternal characteristics
Social charasteristics
Involves variables 8-17}
\usage{
du.quality.local.core.maternal.social.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of maternal characteristics
Social charasteristics
Involves variables 8-17
}
\keyword{internal}
