% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.meta}
\alias{du.quality.local.core.meta}
\title{Run checks on metadata variables in LifeCycle
It involves variables 1-7}
\usage{
du.quality.local.core.meta(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Run checks on metadata variables in LifeCycle
It involves variables 1-7
}
\keyword{internal}
