% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.paternal.health.related.char}
\alias{du.quality.local.core.paternal.health.related.char}
\title{Part of paternal characteristics
Paternal health-related characteristics
Involves variables 68-76}
\usage{
du.quality.local.core.paternal.health.related.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of paternal characteristics
Paternal health-related characteristics
Involves variables 68-76
}
\keyword{internal}
