% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.paternal.lifestyle.char}
\alias{du.quality.local.core.paternal.lifestyle.char}
\title{Part of paternal characteristics
Paternal lifestyle characteristics
Involves variables 77-78}
\usage{
du.quality.local.core.paternal.lifestyle.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of paternal characteristics
Paternal lifestyle characteristics
Involves variables 77-78
}
\keyword{internal}
