% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core.paternal.socio.demo.char}
\alias{du.quality.local.core.paternal.socio.demo.char}
\title{Part of paternal characteristics
Paternal socio-demographic characteristics
Involves variables 46-67}
\usage{
du.quality.local.core.paternal.socio.demo.char(lc_data)
}
\arguments{
\item{lc_data}{core variables from LifeCycle}
}
\description{
Part of paternal characteristics
Paternal socio-demographic characteristics
Involves variables 46-67
}
\section{Warning}{

\enumerate{
  \item 46 - occup_f1_?

        If a cohort does not have data specifying whether the father is employed or self-employed, categorise the father as employed; the variable will be partially harmonised!. If variable is partially harmonised, please check you have documented the reason for this accordingly in detail in the Online Catalogue under the 'match'. Check for correct type [categorical], number [six] and definition of categories [1=Empl, 2=Self-empl, 3=Unempl, 4=Stud, 5=Domest, 6=Inactive].Also, check for correct number of time bands [eightteen, 0-17]
  \item 47 - occup_f1_fath?

        Check for correct type [categorical], number [four] and definition of categories [Type of father; 1=Biological father, 2=Social father, 3=Social mother, 4=Unknown]. Also, check for correct number of time bands [eightteen, 0-17]
}
}

\keyword{internal}
