% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_local_core.R
\name{du.quality.local.core}
\alias{du.quality.local.core}
\title{Quality control of core harmonized variables in LifeCycle}
\usage{
du.quality.local.core(
  input_format = "CSV",
  input_path,
  variable_category = "ALL"
)
}
\arguments{
\item{input_format}{possible formats are CSV,STATA,SPSS or SAS (default = CSV)}

\item{input_path}{path for importfile}

\item{variable_category}{describes the categories you want to run in your quality control (possible values are: ALL, META, MATERNAL, PATERNAL, CHILD and HOUSEHOLD - default = ALL)}
}
\description{
Quality control of core harmonized variables in LifeCycle
}
\details{
Procedures for quality checks of harmonised core variables
\enumerate{
  \item Verify the list of variables and formats
  \item Check the univariate distributions
  \item Check internal validity
  \item Check consistency in repeated measures
}

You can run all of the categories or seperate ones
\itemize{
  \item All catagories (ALL)
  \item Meta variables (META)
  \item Maternal characteristics (MATERNAL)
  \item Paternal characteristics (PATERNAL)
  \item Child characteristics (CHILD)
  \item Househild characteristics (HOUSEHOLD)
}
}
\section{Warning}{

Some of the functions here will throw "Error's". This is most likely
due to the fact that the variable that you are cross checking has no values and only missing
If that is the case, just make sure the variable exists in your data set and move on to the
next.
}

\keyword{internal}
