% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_helpers.R
\name{du.read.source.file}
\alias{du.read.source.file}
\title{Read the input file from different sources}
\usage{
du.read.source.file(input_path, input_format)
}
\arguments{
\item{input_path}{path for importfile}

\item{input_format}{possible formats are CSV,STATA,SPSS or SAS (default = CSV)}
}
\value{
dataframe with source data
}
\description{
Read the input file from different sources
}
\keyword{internal}
