% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_beta.R
\name{du.reshape.beta}
\alias{du.reshape.beta}
\title{Reshape Script for R: LifeCycle Harmonized Data}
\usage{
du.reshape.beta(
  upload,
  data_version,
  input_format,
  dict_name,
  input_path,
  non_interactive
)
}
\arguments{
\item{upload}{do you want automatically upload the files to your opal (default = true)}

\item{data_version}{version of the data you are going to upload into Opal}

\item{input_format}{possible formats are CSV,STATA,SPSS or SAS (default = CSV)}

\item{dict_name}{version of the dictionary}

\item{non_interactive}{if set to TRUE you will get no questions}

\item{data_input_path}{path for import file}
}
\description{
Reshape Script for R: LifeCycle Harmonized Data
}
\keyword{internal}
