% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{du.reshape.generate.non.repeated}
\alias{du.reshape.generate.non.repeated}
\title{Generate the yearly repeated measures file and write it to your local workspace}
\usage{
du.reshape.generate.non.repeated(
  lc_data,
  upload_to_opal,
  output_path,
  file_prefix,
  dict_kind,
  file_version,
  file_name
)
}
\arguments{
\item{lc_data}{data frame with all the data based upon the CSV file}

\item{upload_to_opal}{do you want to upload to Opal (default = true)}

\item{output_path}{directory where the CSV files need to be stored}

\item{file_prefix}{the date of the generated file}

\item{dict_kind}{can be 'core' or 'outcome'}

\item{file_version}{version of the data release (e.g. 1_0)}

\item{file_name}{non-repeated, monthly-repeated or yearly-repeated}
}
\description{
Generate the yearly repeated measures file and write it to your local workspace
}
