% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_helpers.R
\name{du.reshape.generate.trimesterly.repeated}
\alias{du.reshape.generate.trimesterly.repeated}
\title{Generate the trimesterly repeated measures file and write it to your local workspace}
\usage{
du.reshape.generate.trimesterly.repeated(data, dict_kind)
}
\arguments{
\item{data}{data frame with all the data based upon the CSV file}

\item{dict_kind}{can be 'core' or 'outcome'}
}
\description{
Generate the trimesterly repeated measures file and write it to your local workspace
}
\keyword{internal}
