% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_reshape.R
\name{du.reshape.methyl.clocks}
\alias{du.reshape.methyl.clocks}
\title{Reshape Script for R: LifeCycle Harmonized Data}
\usage{
du.reshape.methyl.clocks(
  upload,
  data_version,
  dict_name,
  data,
  non_interactive
)
}
\arguments{
\item{upload}{do you want automatically upload the files to your opal (default = true)}

\item{data_version}{version of the data you are going to upload into Opal}

\item{dict_name}{version of the dictionary}

\item{data}{methylation data}

\item{non_interactive}{if set to TRUE you will get no questions}
}
\description{
Reshape Script for R: LifeCycle Harmonized Data
}
\keyword{internal}
