% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{du.summarize}
\alias{du.summarize}
\title{This function creates a summary table}
\usage{
du.summarize(df, .var)
}
\arguments{
\item{df}{data frame to summarise}

\item{.var}{variable to summarise}
}
\value{
a summary of the data
}
\description{
This function creates a summary table
}
\keyword{internal}
