% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_methyl.R
\name{du.upload.methyl.clocks}
\alias{du.upload.methyl.clocks}
\title{Generate clocks for uploading to Opal}
\usage{
du.upload.methyl.clocks(
  upload = TRUE,
  dict_name = "",
  action = du.enum.action()$ALL,
  data_input_path = "",
  data_version = "1_0",
  age_when_measured,
  database_name = "opal_data"
)
}
\arguments{
\item{upload}{do we need to upload the DataSHIELD backend}

\item{dict_name}{name of the dictionary located on Github usually something like this: diabetes/test_vars_01}

\item{action}{action to be performed, can be 'reshape', 'populate' or 'all'}

\item{data_input_path}{path to the methylation data}

\item{data_version}{version of the raw data}

\item{age_when_measured}{age of the child when the measurement took place (in months)}

\item{database_name}{is the name of the data backend of DataSHIELD, default = opal_data}
}
\description{
Generate clocks for uploading to Opal
}
\examples{

\dontrun{
du.upload.methyl.clocks(
  dict_name = "methylation_clocks",
  data_input_path = "https://github.com/isglobal-brge/methylclock/blob/master/inst/extdata/MethylationDataExample55.csv?raw=true",
  age_when_measured = 44
)
}

}
