% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{du.upload}
\alias{du.upload}
\title{Upload dictionaries and data into the DataSHIELD backend.}
\usage{
du.upload(
  dict_version,
  data_version = "1_0",
  dict_kind,
  cohort_id,
  database_name = "opal_data",
  data_input_format = du.enum.input.format()$CSV,
  data_input_path,
  action = du.enum.action()$ALL,
  upload = TRUE,
  run_mode = du.enum.run.mode()$NORMAL,
  override_project = NULL,
  download_method = "libcurl"
)
}
\arguments{
\item{dict_version}{version of the data dictionary to be used}

\item{data_version}{version of the dataset to be uploaded}

\item{dict_kind}{can be 'core' or 'outcome'}

\item{cohort_id}{cohort name: [du.enum.cohorts]}

\item{database_name}{is the name of the data backend of DataSHIELD, default = opal_data}

\item{data_input_format}{format of the database to be reshaped. Can be [du.enum.input.format]}

\item{data_input_path}{path to the to-be-uploaded data}

\item{action}{action to be performed, can be [du.enum.action]}

\item{upload}{directly upload the reshaped database to the logged in DataSHIELD server, TRUE or FALSE}

\item{run_mode}{default = NORMAL, [du.enum.run.mode]}

\item{override_project}{overrides the generated project name}

\item{download_method}{method to be passed to `method` argument of download.file. Default is "libcurl".}
}
\description{
Upload dictionaries and data into the DataSHIELD backend.
}
\examples{
\dontrun{
du.upload(
  data_input_format = "CSV",
  data_input_path = "~/path-to-file/all_measurements_v1_2.csv",
  data_version = "1_0",
  dict_version = "2_1",
  cohort_id = "gecko"
)
}

}
