% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc.non.repeated}
\alias{qc.non.repeated}
\title{Check non repeated measures}
\usage{
qc.non.repeated(conns, table, verbose)
}
\arguments{
\item{conns}{connections object for DataSHIELD backend}

\item{table}{to quality check}

\item{verbose}{print verbose output}
}
\description{
Check non repeated measures
}
\keyword{internal}
