% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc.yearly.repeated}
\alias{qc.yearly.repeated}
\title{Quality control for yearly repeated measures}
\usage{
qc.yearly.repeated(conns, table, verbose)
}
\arguments{
\item{conns}{connection object for DataSHIELD backends}

\item{table}{table to perform quality control on}

\item{verbose}{print output to screen}
}
\description{
Quality control for yearly repeated measures
}
\keyword{internal}
