% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_armadillo.R
\name{du.armadillo.import}
\alias{du.armadillo.import}
\title{Import data into Armadillo}
\usage{
du.armadillo.import(project, file_name)
}
\arguments{
\item{project}{the target project where the data should be put into}

\item{file_name}{file_name that should be loaded to generate the data from}
}
\description{
Import data into Armadillo
}
\keyword{internal}
