% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{du.dict.download}
\alias{du.dict.download}
\title{Download all released data dictionaries}
\usage{
du.dict.download(dict_name, dict_version, dict_kind)
}
\arguments{
\item{dict_name}{used for beta dictionaries}

\item{dict_version}{dictionary version}

\item{dict_kind}{dictionary kind (possible kinds are 'core' or 'outcome')}
}
\description{
Download all released data dictionaries
}
\keyword{internal}
