% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_helpers.R
\name{du.dict.import}
\alias{du.dict.import}
\title{Import the data dictionaries into Opal}
\usage{
du.dict.import(project, dict_version, dict_kind, data_version)
}
\arguments{
\item{project}{project in which the data is imported}

\item{dict_version}{dictionary version (concerning core or outcome)}

\item{dict_kind}{dictionary kind (core or outcome)}

\item{data_version}{version of the data (specific to the cohort)}
}
\description{
Import the data dictionaries into Opal
}
\keyword{internal}
