% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{du.login}
\alias{du.login}
\title{Login into the opal instance and}
\usage{
du.login(hostname, username = "administrator", password, insecure = FALSE)
}
\arguments{
\item{hostname}{specify hostname of Opal instance}

\item{username}{specify username (of administrator) of Opal instance (default = administrator)}

\item{password}{specify password (of administrator) of Opal instance}

\item{insecure}{NOT RECOMMENDED, use an insecure SSL connection}
}
\description{
Login into the opal instance and
}
\examples{
\dontrun{
du.login(
  hostname = "https://cohort-datashield-server.org",
  username = "local_dm",
  password = "very-secret-pw"
)
}

}
