% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{du.match.dict.categories}
\alias{du.match.dict.categories}
\title{Match the categories with the variables to be import them
Import the variables}
\usage{
du.match.dict.categories(project, dict_kind, table, file_name)
}
\arguments{
\item{project}{project resource in Opal}

\item{dict_kind}{kind of dictionary}

\item{table}{table name}

\item{file_name}{dictionary to upload to Opal}
}
\description{
Match the categories with the variables to be import them
Import the variables
}
\keyword{internal}
