% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{du.quality.control}
\alias{du.quality.control}
\title{Validates the variables for a certain table}
\usage{
du.quality.control(project, verbose = FALSE)
}
\arguments{
\item{project}{specify project you want to perform quality control on}

\item{verbose}{ouput the functions output when set to TRUE}
}
\description{
Validates the variables for a certain table
}
