% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{du.reshape}
\alias{du.reshape}
\title{Reshape Script for R: LifeCycle Harmonized Data}
\usage{
du.reshape(
  upload = TRUE,
  data_version,
  input_format,
  dict_version,
  dict_kind,
  input_path,
  non_interactive
)
}
\arguments{
\item{upload}{do you want automatically upload the files to your opal (default = true)}

\item{data_version}{version of the data you are going to upload into Opal}

\item{input_format}{possible formats are CSV,STATA,SPSS or SAS (default = CSV)}

\item{dict_version}{version of the dictionary}

\item{dict_kind}{kind of data to reshape (default = core)}

\item{input_path}{path for import file}

\item{non_interactive}{if set to TRUE you will get no questions}
}
\description{
Reshape Script for R: LifeCycle Harmonized Data
}
\keyword{internal}
