% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{du.retrieve.dictionaries}
\alias{du.retrieve.dictionaries}
\title{Retrieve the right file from download directory}
\usage{
du.retrieve.dictionaries(dict_table, dict_kind)
}
\arguments{
\item{dict_table}{a specific table that you want to check}

\item{dict_kind}{can be 'core' or 'outcome'}
}
\value{
a raw version of the dictionary
}
\description{
Retrieve the right file from download directory
}
\keyword{internal}
