% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_beta.R
\name{du.upload.beta}
\alias{du.upload.beta}
\title{Upload beta dictionaries to your Opal instance}
\usage{
du.upload.beta(
  upload = TRUE,
  dict_name = "",
  action = du.enum.action()$ALL,
  data_input_path = "",
  data_input_format = du.enum.input.format()$CSV,
  database_name = "opal_data"
)
}
\arguments{
\item{upload}{do we need to upload the DataSHIELD backend}

\item{dict_name}{name of the dictionary located on Github usually something like this: diabetes/test_vars_01}

\item{action}{action to be performed, can be 'reshape', 'populate' or 'all'}

\item{data_input_path}{path to the to-be-reshaped data}

\item{data_input_format}{format of the database to be reshaped. Can be 'CSV', 'STATA', or 'SAS'}

\item{database_name}{is the name of the data backend of DataSHIELD, default = opal_data}
}
\description{
Upload beta dictionaries to your Opal instance
}
