% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{du.upload}
\alias{du.upload}
\title{Upload dictionaries and data into the DataSHIELD backend.}
\usage{
du.upload(
  dict_version = "2_1",
  data_version = "1_1",
  dict_kind = "core",
  cohort_id,
  database_name = "opal_data",
  data_input_format = "CSV",
  data_input_path,
  data_output_path = getwd(),
  action = "all",
  upload_to_opal = TRUE,
  non_interactive = FALSE
)
}
\arguments{
\item{dict_version}{version of the data dictionnary to be used}

\item{data_version}{version of the dataset to be uploaded}

\item{dict_kind}{can be 'core' or 'outcome'}

\item{cohort_id}{cohort name}

\item{database_name}{is the name of the data backend of DataSHIELD backend: 'Opal'}

\item{data_input_format}{format of the database to be reshaped. Can be 'CSV', 'STATA', or 'SAS'}

\item{data_input_path}{Path to the to-be-reshaped database}

\item{data_output_path}{Path where the reshaped databases will be written}

\item{action}{action to be performed, can be 'reshape', 'populate' or 'all'}

\item{upload_to_opal}{Wether to directly upload the reshaped database to the logged in opal server}

\item{non_interactive}{if set to TRUE you will get no questions}
}
\description{
Upload dictionaries and data into the DataSHIELD backend.
}
\examples{
\dontrun{
du.upload(
  data_input_format = "CSV",
  data_input_path = "~/path-to-file/all_measurements_v1_2.csv", 
  data_version = "1_0",
  dict_version = "2_1",
  cohort_id = "gecko"
)
}

}
